fix: Move authOptions to separate file for Next.js App Router compliance
- Create src/lib/auth/options.ts with authOptions export
- Update route.ts to import from the new location
- Update admin.ts import path

Next.js App Router doesn't allow exporting non-route fields from route files.
